src/app/quiz/keyboard/keyboard.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { KeyboardComponent } from './keyboard.component';

describe('KeyboardComponent', () => {
  let component: KeyboardComponent;
  let fixture: ComponentFixture<KeyboardComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [KeyboardComponent],
    }).compileComponents();

    fixture = TestBed.createComponent(KeyboardComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render 26 buttons', () => {
    const buttons = fixture.nativeElement.querySelectorAll('button');
    expect(buttons.length).toBe(26);
  });

  it('button have a click event handler', () => {
    const clickSpy = spyOn(component, 'click');
    const button = (fixture.nativeElement as HTMLElement).querySelector('button')!;
    button.click();
    expect(clickSpy).toHaveBeenCalled();
  });
});
